+++ src/main/java/com/lescastcodeurs/bot/SlackBotAction.java
@@ -95,7 +95,8 @@ public enum SlackBotAction {
        • La formatage suivant est conservé : *gras*, _italique_, ~barré~, `code`. Les citations sont aussi possibles, mais uniquement sur le premier message du thread.
        • Les retours chariot ne sont conservés que sur le premier message du thread, ou quand les réponses sont organisées sous la forme d'une liste.
        • Les show notes peuvent être publiées plusieurs fois: le fichier markdown est alors mis à jour.
-       • La liste des conférences françaises disponible sur <https://github.com/scraly/developers-conferences-agenda> est automatiquement récupérée pour compléter la section _Conférences_.
+       • La liste des conférences françaises disponible est automatiquement récupérée pour compléter la section _Conférences_. Voir le README du projet pour plus d'informations sur le fonctionnement de cette fonctionnalité.
+       • La date d'enregistrement du podcast est automatiquement devinée à partir des premiers messages des threads. Voir le README du projet pour plus d'informations sur le fonctionnement de cette fonctionnalité.
       """,
       Constants.GENERATE_SHOW_NOTES_ADDRESS),